#78 [refactor] result 어그리게이트의 dto를 응용계층으로 이동한다.:
