+++ src/main/java/kusitms/backend/result/application/ResultService.java
@@ -12,10 +12,10 @@ import kusitms.backend.result.domain.enums.KtWizStadiumStatusType;
 import kusitms.backend.result.domain.enums.ProfileStatusType;
 import kusitms.backend.result.domain.enums.StadiumStatusType;
 import kusitms.backend.result.domain.repository.ResultRepository;
-import kusitms.backend.result.dto.request.SaveTopRankedZoneRequestDto;
-import kusitms.backend.result.dto.response.GetProfileResponseDto;
-import kusitms.backend.result.dto.response.GetZonesResponseDto;
-import kusitms.backend.result.dto.response.SaveTopRankedZoneResponseDto;
+import kusitms.backend.result.application.dto.request.SaveTopRankedZoneRequestDto;
+import kusitms.backend.result.application.dto.response.GetProfileResponseDto;
+import kusitms.backend.result.application.dto.response.GetZonesResponseDto;
+import kusitms.backend.result.application.dto.response.SaveTopRankedZoneResponseDto;
 import kusitms.backend.result.status.ResultErrorStatus;
 import kusitms.backend.stadium.application.StadiumService;
 import kusitms.backend.stadium.status.StadiumErrorStatus;
