+++ src/main/java/kusitms/backend/result/application/dto/response/SaveTopRankedZoneResponseDto.java
@@ -1,4 +1,4 @@
-package kusitms.backend.result.dto.response;
+package kusitms.backend.result.application.dto.response;
 
 import lombok.Builder;
