+++ src/main/java/kusitms/backend/result/presentation/ResultController.java
@@ -4,10 +4,10 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import kusitms.backend.global.dto.ApiResponse;
 import kusitms.backend.result.application.ResultService;
-import kusitms.backend.result.dto.request.SaveTopRankedZoneRequestDto;
-import kusitms.backend.result.dto.response.GetProfileResponseDto;
-import kusitms.backend.result.dto.response.GetZonesResponseDto;
-import kusitms.backend.result.dto.response.SaveTopRankedZoneResponseDto;
+import kusitms.backend.result.application.dto.request.SaveTopRankedZoneRequestDto;
+import kusitms.backend.result.application.dto.response.GetProfileResponseDto;
+import kusitms.backend.result.application.dto.response.GetZonesResponseDto;
+import kusitms.backend.result.application.dto.response.SaveTopRankedZoneResponseDto;
 import kusitms.backend.result.status.ResultSuccessStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
