+++ src/test/java/kusitms/backend/result/ResultControllerTest.java
@@ -8,10 +8,10 @@ import kusitms.backend.configuration.ControllerTestConfig;
 import kusitms.backend.result.application.ResultService;
 import kusitms.backend.result.common.Reference;
 import kusitms.backend.result.common.ReferencesGroup;
-import kusitms.backend.result.dto.request.SaveTopRankedZoneRequestDto;
-import kusitms.backend.result.dto.response.GetProfileResponseDto;
-import kusitms.backend.result.dto.response.GetZonesResponseDto;
-import kusitms.backend.result.dto.response.SaveTopRankedZoneResponseDto;
+import kusitms.backend.result.application.dto.request.SaveTopRankedZoneRequestDto;
+import kusitms.backend.result.application.dto.response.GetProfileResponseDto;
+import kusitms.backend.result.application.dto.response.GetZonesResponseDto;
+import kusitms.backend.result.application.dto.response.SaveTopRankedZoneResponseDto;
 import kusitms.backend.result.presentation.ResultController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
